Moved removed references to sumi::transport.
This is the start of a large refactor in which sumi is going to be
consolidated into a single library instead of two versions.  I also
moved sumi::message into a deprecated namespace with the intent of
eventually removing it.

A large number of files were touched in this commit because all
references to ::sumi::transport have now been switch to
::sstmac::sumi::transport.  The next step is to move all of the code
from ::sumi::transport into ::sstmac::sumi::transport and remove the
::sumi::transport class entirely.